test: rewrite SonarQube verification to match rule's documented shape
Previous version (string concatenated into a variable, then passed to
execute()) produced 0 issues and 0 hotspots on SonarCloud - likely
because S2077's pattern matching looks for formatting done directly
inside the .execute() call. Rewrote as an f-string inline in the call
(the exact shape Sonar's own rule description shows) and added back a
hardcoded credential (S2068) as a second, more reliably-detected
pattern. Also renamed away from *_test.py in case that suffix gets
treated as test code rather than main code.
